Did a near-complete revamp.
The project uses Python 3.4 now.
Skorbela-server now depends on Autobahn.ws and asyncio.
No more page-updater, asyncio enabled us to read
files while serving. Database is handled by jsondatabase
currently, this might change later.

Websocket connections are no longer single-use:
when a socket is connected, it is assigned an unique ID
and remains open as long as the page is open. If a change
is made, the change is emitted to all open websockets.